feat: update type models

+++ backend/src/main/java/com/example/backend/model/Seat.java
@@ -8,8 +8,8 @@ import java.util.Objects;
 
 @Hidden
 public class Seat {
-    private int id;
-    private int cinemaHallId;
+    private Long id;
+    private Long cinemaHallId;
     private String seatNumber;
     private BigDecimal price;
     private SeatStatus status;
@@ -28,7 +28,7 @@ public class Seat {
 
     }
 
-    public Seat(int id, int cinemaHallId, String seatNumber, BigDecimal price, SeatStatus status) {
+    public Seat(Long id, Long cinemaHallId, String seatNumber, BigDecimal price, SeatStatus status) {
         super();
         this.id = id;
         this.cinemaHallId = cinemaHallId;
@@ -37,19 +37,19 @@ public class Seat {
         this.status = status;
     }
 
-    public int getId() {
+    public Long getId() {
         return id;
     }
 
-    public void setId(int id) {
+    public void setId(Long id) {
         this.id = id;
     }
 
-    public int getCinemaHallId() {
+    public Long getCinemaHallId() {
         return cinemaHallId;
     }
 
-    public void setCinemaHallId(int cinemaHallId) {
+    public void setCinemaHallId(Long cinemaHallId) {
         this.cinemaHallId = cinemaHallId;
     }
